backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# ML & Data
xgboost==2.0.3
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
shap==0.44.1
h3==4.0.0b5
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
pyyaml==6.0.1

# Date/Time
python-dateutil==2.8.2
holidays==0.39

# CORS
fastapi-cors==0.0.6
